requirements.txt
PyQt5~=5.15.2
ndu-gate-camera
PyYAML~=5.3.1
tendo~=0.2.15
opencv-python